Bug 8518 partially fixed: Error messages when exporting were not clear

+++ scilab/modules/graphic_export/src/java/org/scilab/modules/graphic_export/FileExporter.java
@@ -141,7 +141,7 @@ public class FileExporter {
         //Put back the old infoMessage
         exportedFig.setInfoMessage(oldInfoMessage);
 
-        if (saveFileType != -1) {
+        if (saveFileType != -1 && ExportRenderer.getErrorNumber() == ExportRenderer.SUCCESS) {
             ConvertSVG.SVGTo(fileName, saveFileName, saveFileType);
             new File(fileName).delete();
         }
